[ci skip] Fixed garray_T member types
This patch fixes the types of the members in garray_T:

    * int ga_len -> unsigned int
    * int ga_maxlen -> unsigned int
    * int ga_itemsize -> size_t
    * int ga_growsize -> unsigned int

And patches the function prototypes (but without changing the number or
order of the) arguments therefor.

+++ src/garray.h
@@ -7,19 +7,19 @@
 /// This is used to store information that only grows, is deleted all at
 /// once, and needs to be accessed by index.  See ga_clear() and ga_grow().
 typedef struct growarray {
-  int ga_len;                       // current number of items used
-  int ga_maxlen;                    // maximum number of items possible
-  int ga_itemsize;                  // sizeof(item)
-  int ga_growsize;                  // number of items to grow each time
-  void *ga_data;                    // pointer to the first item
+  unsigned int  ga_len;         // current number of items used
+  unsigned int  ga_maxlen;      // maximum number of items possible
+  size_t        ga_itemsize;    // sizeof(item)
+  unsigned int  ga_growsize;    // number of items to grow each time
+  void          *ga_data;       // pointer to the first item
 } garray_T;
 
 #define GA_EMPTY { 0, 0, 0, 0, NULL }
 
 void ga_clear(garray_T *gap);
 void ga_clear_strings(garray_T *gap);
-void ga_init(garray_T *gap, int itemsize, int growsize);
-void ga_grow(garray_T *gap, int n);
+void ga_init(garray_T *gap, size_t itemsize, unsigned int growsize);
+void ga_grow(garray_T *gap, unsigned int n);
 char_u *ga_concat_strings(garray_T *gap) FUNC_ATTR_NONNULL_RET;
 void ga_remove_duplicate_strings(garray_T *gap);
 void ga_concat(garray_T *gap, char_u *s);
